% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyser.R
\name{analyzer}
\alias{analyzer}
\title{OLATTestAnalyzeR}
\usage{
analyzer(file, sep = ";", encoding = "ISO-8859-1")
}
\arguments{
\item{file}{xls or csv file}

\item{sep}{tab ('\t') as default delimiter}

\item{encoding}{'ISO-8859-1' as default}
}
\description{
Formats and analysis OLAT - Online Learning and Training Single Choice Tests
}
\details{
Features:
  - Formatted output with ...
    - letters as column names
    - marks (*) the correct answers
    - Sum of answer columns
    - Mean Test score
    - Proportion of correct answers (p = difficulty)
    - Proportion of incorrect answers
    - Point-biserial correlation coefficient for each question (r = selectivity)
}
\examples{
 test <- analyser('OLAT SCQs.xls', sep = '\\t') # reading xls-file with analyzer()
 
}
\author{
Benjamin Gerwoll-Ronca
}
\keyword{Item}
\keyword{analysis}

