% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writer.R
\name{analyzer.write}
\alias{analyzer.write}
\title{OLATTestAnalyzeR}
\usage{
analyzer.write(df, file = "OLAT_Test_Analysis.xls", keys = NULL,
  sep = "\\t", na = "", row.names = FALSE)
}
\arguments{
\item{df}{output of the analyzer() function}

\item{file}{(path and) name of the output file}

\item{keys}{csv file (2 column) with key of the correct answer}

\item{sep}{tab ('\t') as default delimiter}
}
\description{
Writes formatted output of the analyser() function to a xls or csv file
}
\examples{
analyser.write(test, 'OLAT SCQs formatted.xls', keys = 'question_keys.csv') # writing formatted test into xls-file

}
\author{
Benjamin Gerwoll-Ronca
}
\keyword{OLAT}
\keyword{analysis}}
\keyword{{Item}

