% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_discrimination.R, R/plot_score.R,
%   R/plots.R
\name{plot_discrimination}
\alias{plot_discrimination}
\title{OLATTestAnalyzeR}
\usage{
plot_discrimination(df)

plot_discrimination(df)

plot_discrimination(df)
}
\arguments{
\item{df}{output of analyzer() function}

\item{df}{output of analyzer() function}

\item{df}{output of analyzer() function}
}
\description{
Scatter plot Difficulty vs. Selectivity

Bar chart Distribution of Test score

Formats and analysis OLAT - Online Learning and Training Single Choice Tests
}
\details{
Features:
  - Scatter plot Difficulty vs. Selectivity
  - Bar chart Distribution of Test score
  - Bar chart Distribution of completed exams over examination period
}
\examples{
 plot_discrimination(test) # scatter plot of selectivity vs. difficulty
 
 plot_score(test)  # bar chart distribution of test scores
 
 plot_discrimination(test)                     # scatter plot of selectivity vs. difficulty
 
 plot_score(test)                              # bar chart distribution of test scores
 
}
\author{
Benjamin Gerwoll-Ronca

Benjamin Gerwoll-Ronca

Benjamin Gerwoll-Ronca
}
\keyword{Item}
\keyword{analysis}

