% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OLATTestAnalyseR.R
\name{analyzer}
\alias{analyzer}
\title{OLATTestAnalyzeR
  
  @author Benjamin Gerwoll-Ronca
  @Version 0.0.9000
  @License GPLv3
  => FOR SINGLE CHOICE QUESTIONS ONLY !
  Features:
  - Formated output with ...
    - Sum of answer columns
    - Mean Test score
    - Proportion of correct answers (p = difficulty)
    - Proportion of incorrect answers
    - Point-biserial correlation coefficient for each question (r = selectivity) 
  - Scatter plot Difficulty vs. Selectivity
  - Bar chart Distribution of Test score
  - Bar chart Distribution of completed exams over examination period
 @export
 @examples
 analyser()
 analyser.write()
 plot_discrimination()
 plot_score()}
\usage{
analyzer(file, sep = ";", encoding = "Latin-1")
}
\description{
OLATTestAnalyzeR
  
  @author Benjamin Gerwoll-Ronca
  @Version 0.0.9000
  @License GPLv3
  => FOR SINGLE CHOICE QUESTIONS ONLY !
  Features:
  - Formated output with ...
    - Sum of answer columns
    - Mean Test score
    - Proportion of correct answers (p = difficulty)
    - Proportion of incorrect answers
    - Point-biserial correlation coefficient for each question (r = selectivity) 
  - Scatter plot Difficulty vs. Selectivity
  - Bar chart Distribution of Test score
  - Bar chart Distribution of completed exams over examination period
 @export
 @examples
 analyser()
 analyser.write()
 plot_discrimination()
 plot_score()
}

