% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OLATTestAnalyseR.R, R/analyser.R
\name{analyzer}
\alias{analyzer}
\title{OLATTestAnalyzeR
  
  @description Formats and analysis OLAT - Online Learning and Training Single Choice Tests
  
  @details 
  Features:
  - Formatted output with ...
    - letters as column names
    - marks (*) the correct answers
    - Sum of answer columns
    - Mean Test score
    - Proportion of correct answers (p = difficulty)
    - Proportion of incorrect answers
    - Point-biserial correlation coefficient for each question (r = selectivity) 
  - Scatter plot Difficulty vs. Selectivity
  - Bar chart Distribution of Test score
  - Bar chart Distribution of completed exams over examination period
 
 @param file xls or csv file
 @param sep tab ('\t') as default delimiter
 @param encoding 'ISO-8859-1' as default
 
 @author Benjamin Gerwoll-Ronca
 @version 0.0.9001
 @license GPL-3
 @keywords Item analysis}
\usage{
analyzer(file, sep = ";", encoding = "ISO-8859-1")

analyzer(file, sep = ";", encoding = "ISO-8859-1")
}
\arguments{
\item{file}{xls or csv file}

\item{sep}{tab ('\t') as default delimiter}

\item{encoding}{'ISO-8859-1' as default}
}
\description{
@examples
 test <- analyser('OLAT SCQs.xls', sep = '\t') # reading xls-file with analyzer()
 
 plot_discrimination(test)                     # scatter plot of selectivity vs. difficulty
 
 plot_score(test)                              # bar chart distribution of test scores
 
 analyser.write(test, 'OLAT SCQs formatted.xls', keys = 'question_keys.csv') # writing formatted test into xls-file

Formats and analysis OLAT - Online Learning and Training Single Choice Tests
}
\details{
@export

Features:
  - Formatted output with ...
    - letters as column names
    - marks (*) the correct answers
    - Sum of answer columns
    - Mean Test score
    - Proportion of correct answers (p = difficulty)
    - Proportion of incorrect answers
    - Point-biserial correlation coefficient for each question (r = selectivity)
}
\examples{
 test <- analyser('OLAT SCQs.xls', sep = '\\t') # reading xls-file with analyzer()
 
}
\author{
Benjamin Gerwoll-Ronca
}
\keyword{Item}
\keyword{analysis}

